requirements.txt
google-adk
litellm
pre-commit
autogen-agentchat
autogen-ext[anthropic]
python-dotenv
praw
asyncpraw
nest-asynchio